Revert "corrections"
This reverts commit 9912400ac39c03b0cb96205103c59e196554d19f.

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -8,7 +8,7 @@ from typing import List
 wait_ran = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: int, max_delay: int) -> List[float]:
+async def wait_n(n: int, max_delay: int) -> List[int]:
     '''returns a list of the outputs of wait_random'''
     res = await asyncio.gather(
         *tuple(map(lambda _: wait_ran(max_delay), range(n)))
